seperate Search component
